switch to phasor & jitter testing
Switched from metro to phasor-based timing.  This resolved the jitter/desychronization we were seeing on the max side.  phasor timing is audio-sample-rate precise, whereas metro timing is only precise to ~1ms.

This version also has hooks to disable wifi and ISRs, and stores jitter in a 10000-sample array.  When the array is full it stops the show and sends a message over UDP.  That data can be recovered by sending 'J' from MAX over UDP to write into a [coll] object for data analysis.